feat(core): restructure text animation tests into module
